Update_Pipeline_problematicCol_imputationCol_Enc
The file was improved by adding an imputation for every column( does not matter if categorical or numerical), processing of problematic columns (columns with unique values for each cells which are not appropriate for ML) - dropping or encoding (ordinal and one-hot encoding), encoding of labels (label binarizer) and pipeline. More complex tasks have to be solved - Runtime of app, testing of binary classification, xx-samples for classification models etc.